+++ openair1/PHY/NR_TRANSPORT/nr_ulsch_decoding.c
@@ -94,7 +94,7 @@ int nr_ulsch_decoding(PHY_VARS_gNB *phy_vars_gNB,
                       uint32_t frame,
                       uint8_t nr_tti_rx,
                       uint32_t *G,
-                      uint8_t *ULSCH_ids,
+                      int *ULSCH_ids,
                       int nb_pusch)
 {
   nrLDPC_TB_decoding_parameters_t TBs[nb_pusch];
